Added some tests scripts and fixed the underlying bug.

+++ contracts/SampleContract.sol
@@ -3,7 +3,7 @@ pragma solidity ^0.4.4;
 contract SampleContract {
 
     // This value is visible in etherscan.io explorer
-    uint public value;
+    uint256 public value;
 
     // Anyone can call this contract and override the value of the previous caller
     function setValue(uint256 value_) public {
